+++ apps/cowswap-frontend/src/legacy/state/orders/middleware/soundMiddleware.test.ts
@@ -14,7 +14,7 @@ const actionMock = mock<AnyAction>()
 jest.mock('modules/sounds')
 
 // TODO: Break down this large function into smaller functions
- 
+
 describe('soundMiddleware', () => {
   beforeEach(() => {
     resetCalls(actionMock)
@@ -23,8 +23,8 @@ describe('soundMiddleware', () => {
       orders: {
         1: 'mock orders',
       },
-    // TODO: Replace any with proper type definitions
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      // TODO: Replace any with proper type definitions
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
     } as any)
     jest.clearAllMocks()
   })
@@ -35,8 +35,8 @@ describe('soundMiddleware', () => {
       when(actionMock.type).thenReturn('order/fullfillOrdersBatch')
       when(mockStore.getState()).thenReturn({
         orders: {},
-      // TODO: Replace any with proper type definitions
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        // TODO: Replace any with proper type definitions
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
       } as any)
 
       soundMiddleware(instance(mockStore))(nextMock)(instance(actionMock))
@@ -120,8 +120,8 @@ describe('soundMiddleware', () => {
             },
           },
         },
-      // TODO: Replace any with proper type definitions
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        // TODO: Replace any with proper type definitions
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
       } as any)
 
       soundMiddleware(instance(mockStore))(nextMock)(instance(actionMock))
@@ -147,8 +147,8 @@ describe('soundMiddleware', () => {
             },
           },
         },
-      // TODO: Replace any with proper type definitions
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        // TODO: Replace any with proper type definitions
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
       } as any)
 
       soundMiddleware(instance(mockStore))(nextMock)(instance(actionMock))