+++ apps/widget-configurator/src/app/configurator/utils/validateURL.ts
@@ -8,7 +8,7 @@ export const validateURL = (url: string) => {
       '(\\:\\d+)?(\\/[-a-z\\d%_.~+]*)*' + // port and path
       '(\\?[;&a-z\\d%_.~+=-]*)?' + // query string
       '(\\#[-a-z\\d_]*)?$',
-    'i'
+    'i',
   ) // fragment locator
   return !!pattern.test(url)
 }