+++ libs/common-const/src/chainInfo.ts
@@ -11,18 +11,20 @@ import {
   plasma,
   polygon,
   sepolia,
+  ink,
   SupportedChainId,
+  HttpsString,
 } from '@cowprotocol/cow-sdk'
 
 import { NATIVE_CURRENCIES } from './nativeAndWrappedTokens'
 import { TokenWithLogo } from './types'
 
 export interface BaseChainInfo {
-  readonly docs: string
-  readonly bridge?: string
-  readonly explorer: string
-  readonly infoLink: string
-  readonly logo: { light: string; dark: string }
+  readonly docs: HttpsString
+  readonly bridge?: HttpsString
+  readonly explorer: HttpsString
+  readonly infoLink: HttpsString
+  readonly logo: { light: HttpsString; dark: HttpsString }
   readonly name: string
   readonly addressPrefix: string
   readonly label: string
@@ -56,7 +58,10 @@ function mapChainInfoToBaseChainInfo(
     bridge: chainInfo.bridges?.[0]?.url,
     explorer: chainInfo.blockExplorer.url ?? '',
     infoLink: chainInfo.website.url,
-    logo: chainInfo.logo,
+    logo: {
+      light: chainInfo.logo.light as HttpsString,
+      dark: chainInfo.logo.dark as HttpsString,
+    },
     addressPrefix: chainInfo.addressPrefix,
     label: chainInfo.label,
     explorerTitle: chainInfo.blockExplorer.name,
@@ -132,6 +137,12 @@ export const CHAIN_INFO: ChainInfoMap = {
     urlAlias: 'plasma',
     nativeCurrency: NATIVE_CURRENCIES[SupportedChainId.PLASMA],
   },
+  [SupportedChainId.INK]: {
+    ...mapChainInfoToBaseChainInfo(ink),
+    name: 'ink',
+    urlAlias: 'ink',
+    nativeCurrency: NATIVE_CURRENCIES[SupportedChainId.INK],
+  },
   [SupportedChainId.SEPOLIA]: {
     ...mapChainInfoToBaseChainInfo(sepolia),
     name: 'sepolia',
@@ -142,6 +153,7 @@ export const CHAIN_INFO: ChainInfoMap = {
 
 /**
  * Sorted array of chain IDs in order of relevance.
+ * TODO: Sort by TVL? Reference: https://defillama.com/chain/gnosis
  */
 export const SORTED_CHAIN_IDS: SupportedChainId[] = [
   SupportedChainId.MAINNET,
@@ -152,6 +164,7 @@ export const SORTED_CHAIN_IDS: SupportedChainId[] = [
   SupportedChainId.AVALANCHE,
   SupportedChainId.LINEA, // TODO: decide where to place Linea
   SupportedChainId.PLASMA, // TODO: decide where to place Plasma
+  SupportedChainId.INK, // TODO: decide where to place Ink
   SupportedChainId.GNOSIS_CHAIN,
   SupportedChainId.LENS,
   SupportedChainId.SEPOLIA,