+++ libs/wallet/src/web3-react/connectors/metaMaskSdk/index.ts
@@ -10,7 +10,7 @@ import { Connector } from '@web3-react/types'
 
 import { parseChainId } from '../../utils/parseChainId'
 
-import type { MetaMaskSDK as _MetaMaskSDK, MetaMaskSDKOptions as _MetaMaskSDKOptions, SDKProvider } from '@metamask/sdk'
+import type { MetaMaskSDK as _MetaMaskSDK, MetaMaskSDKOptions as _MetaMaskSDKOptions } from '@metamask/sdk'
 
 /**
  * MetaMaskSDK options.
@@ -23,6 +23,10 @@ type MetaMaskSDKOptions = Pick<_MetaMaskSDKOptions, 'infuraAPIKey' | 'readonlyRP
  * Listener type for MetaMaskSDK events.
  */
 type Listener = Parameters<Provider['on']>[1]
+
+type MetaMaskProvider = Provider & {
+  isConnected?: () => boolean
+}
 
 /**
  * Error thrown when the MetaMaskSDK is not installed.
@@ -50,7 +54,7 @@ export interface MetaMaskSDKConstructorArgs {
  */
 export class MetaMaskSDK extends Connector {
   private sdk?: _MetaMaskSDK
-  provider?: SDKProvider = undefined
+  provider?: MetaMaskProvider = undefined
   private readonly options: MetaMaskSDKOptions
   private eagerConnection?: Promise<void>
 
@@ -112,7 +116,7 @@ export class MetaMaskSDK extends Connector {
         await this.sdk.init()
       }
 
-      this.provider = this.sdk.getProvider()!
+      this.provider = this.sdk.getProvider()! as unknown as MetaMaskProvider
 
       this.provider.on('connect', (({ chainId }: ProviderConnectInfo): void => {
         this.actions.update({ chainId: parseChainId(chainId) })
@@ -247,8 +251,8 @@ export class MetaMaskSDK extends Connector {
           },
         },
       })
-      .then((success) => {
-        if (!success) throw new Error('Rejected')
+      .then((success: unknown) => {
+        if (success !== true) throw new Error('Rejected')
         return true
       })
   }
@@ -269,7 +273,7 @@ export class MetaMaskSDK extends Connector {
 
     const chainIdHex = `0x${desiredChainId.toString(16)}`
     this.provider
-      .request<void>({
+      .request({
         method: 'wallet_switchEthereumChain',
         params: [{ chainId: chainIdHex }],
       })
@@ -281,7 +285,7 @@ export class MetaMaskSDK extends Connector {
         if (originalError.code === 4902 && desiredChain !== undefined) {
           if (!this.provider) throw new NoMetaMaskSDKError()
           // if we're here, we can try to add a new network
-          return this.provider.request<void>({
+          return this.provider.request({
             method: 'wallet_addEthereumChain',
             params: [{ ...desiredChain, chainId: chainIdHex }],
           })